+++ pkg/action/uninstall.go
@@ -27,7 +27,7 @@ import (
 	chartutil "helm.sh/helm/v4/pkg/chart/util"
 	"helm.sh/helm/v4/pkg/kube"
 	"helm.sh/helm/v4/pkg/release"
-	"helm.sh/helm/v4/pkg/releaseutil"
+	releaseutil "helm.sh/helm/v4/pkg/release/util"
 	helmtime "helm.sh/helm/v4/pkg/time"
 )
