+++ pkg/release/util/manifest_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package releaseutil // import "helm.sh/helm/v4/pkg/releaseutil"
+package util // import "helm.sh/helm/v4/pkg/release/util"
 
 import (
 	"reflect"
